enabled ext4 journal layer
This is a generic journaling layer for block devices that support both 32-bit and 64-bit block numbers. It is currently used by the ext4dev/ext4 filesystem, but it could also be used to add journal support to other file systems or block devices such as RAID or LVM.

If you are using ext4dev/ext4, you need to say Y here. If you are not using ext4dev/ext4 then you will probably want to say N.

To compile this device as a module, choose M here. The module will be called jbd2. If you are compiling ext4dev/ext4 into the kernel, you cannot compile this code as a module.